LayoutTests/media/mediacapabilities/vp8-expected.txt


PASS VP8 is supported and not powerEfficient for file
